refactor(api): :recycle: features and trpc package consolidated into api package.
