+++ packages/api/src/greeting/index.ts
