+++ packages/api/src/post/post.controller.ts
@@ -1,7 +1,8 @@
 import { Injectable } from "@nestjs/common";
 
+import type { Logger } from "@acme/logging";
 import type { CreatePost } from "@acme/validators";
-import { Logger, LoggerFactory } from "@acme/logging";
+import { LoggerFactory } from "@acme/logging";
 
 import { PostService } from "./post.service";
