+++ packages/api/src/post/post.service.ts
@@ -1,15 +1,10 @@
 import { Injectable } from "@nestjs/common";
 
+import type { Logger } from "@acme/logging";
 import type { CreatePost } from "@acme/validators";
-import { Logger, LoggerFactory } from "@acme/logging";
+import { LoggerFactory } from "@acme/logging";
 
-type Post = {
-  id: number;
-  title: string;
-  content: string;
-};
-
-let posts = [
+const posts = [
   {
     id: 1,
     title: "test title 1",
