--- packages/features/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@acme/features",
-  "version": "0.2.0",
-  "private": true,
-  "author": "Akash <akashdevcc@gmail.com>",
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "clean": "git clean -xdf .turbo node_modules",
-    "build": "tsup src/index.ts --format cjs,esm --dts",
-    "dev": "pnpm build --watch",
-    "test": "echo 'Add test script here'",
-    "lint": "eslint .",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@acme/logging": "workspace:^",
-    "@acme/trpc": "workspace:^",
-    "@acme/validators": "workspace:^",
-    "@nestjs/common": "^10.0.0",
-    "bcryptjs": "^2.4.3",
-    "jsonwebtoken": "^9.0.2",
-    "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@acme/eslint-config": "workspace:^",
-    "@acme/prettier-config": "workspace:^",
-    "@acme/tsconfig": "workspace:^",
-    "@swc/core": "^1.4.12",
-    "@types/bcryptjs": "^2.4.6",
-    "@types/jsonwebtoken": "^9.0.6",
-    "eslint": "^8.56.0",
-    "prettier": "^3.1.1",
-    "tsup": "^8.0.1",
-    "typescript": "^5.3.3"
-  }
-}
