--- packages/features/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./greeting";
-export * from "./post";
