--- packages/features/src/utils/typeHelpers.ts
@@ -1,1 +1,0 @@
-export type ObjectValues<T> = T[keyof T];
