--- packages/features/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@acme/tsconfig/nest.json",
-  "compilerOptions": {
-    "baseUrl": "./src",
-    "rootDir": "./src",
-    "outDir": "./dist"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "build", "dist"]
-}
