--- packages/trpc/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@acme/trpc",
-  "version": "0.2.0",
-  "private": true,
-  "author": "Akash <akashdevcc@gmail.com>",
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "clean": "git clean -xdf .turbo node_modules",
-    "build": "tsup src/index.ts --format cjs,esm --dts",
-    "dev": "pnpm build --watch",
-    "format": "prettier --check . --ignore-path ../../.gitignore",
-    "lint": "eslint .",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@nestjs/common": "^10.0.0",
-    "@acme/logging": "workspace:^",
-    "@trpc/server": "11.0.0-next-beta.236",
-    "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@acme/eslint-config": "workspace:^",
-    "@acme/prettier-config": "workspace:^",
-    "@acme/tsconfig": "workspace:^",
-    "eslint": "^8.56.0",
-    "prettier": "^3.1.1",
-    "tsup": "^8.0.1",
-    "typescript": "^5.3.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "@acme/eslint-config/base"
-    ]
-  },
-  "prettier": "@acme/prettier-config"
-}
